+++ server/src/db/models/note.py
@@ -6,6 +6,7 @@ from peewee import ForeignKeyField, TextField
 from ...api.models.note import ApiNote
 from ..base import BaseDbModel
 from ..typed import SelectSequence
+from .location import Location
 from .room import Room
 from .user import User
 
@@ -17,16 +18,25 @@ if TYPE_CHECKING:
 class Note(BaseDbModel):
     access: SelectSequence["NoteAccess"]
     shapes: SelectSequence["NoteShape"]
+    room_id: int | None
+    location_id: int | None
 
     uuid = cast(str, TextField(primary_key=True))
     creator = ForeignKeyField(User, backref="notes", on_delete="CASCADE")
     title = cast(str, TextField())
     text = cast(str, TextField())
     tags = cast(str | None, TextField(null=True))
-    room = cast(Room, ForeignKeyField(Room, backref="notes", on_delete="CASCADE"))
+    room = cast(
+        Room | None,
+        ForeignKeyField(Room, null=True, backref="notes", on_delete="CASCADE"),
+    )
+    location = cast(
+        Location | None,
+        ForeignKeyField(Location, null=True, backref="notes", on_delete="CASCADE"),
+    )
 
     def __repr__(self):
-        return f"<Note {self.title} {self.room.get_path()} - {self.creator.name}"
+        return f"<Note {self.title} {self.room.get_path() if self.room else ''} - {self.creator.name}"
 
     def as_pydantic(self):
         tags = json.loads(self.tags) if self.tags else []
@@ -39,5 +49,6 @@ class Note(BaseDbModel):
             tags=tags,
             access=access,
             isRoomNote=self.room is not None,
+            location=self.location_id,
             shapes=[s.shape.uuid for s in self.shapes],
         )